refactor: sorry for this massive commit
it's probably easier to read the faythe source code in its entirety,
than to try to get a grasp of this colossal diff.

the main point here is to try to separate responsibilities, mainly:

1) The "issuer"-module should never know anything other than a CertSpec.
A CertSpec is an abstract definition of a wanted certificate, complete
with CN, SANS and an attached PersistSpec that tells the issuer
where/how to persist the certificate and private key once issued.

2) The "monitor"-module does nothing more than monitor an datasource
   (only kube is implemented still) and create CertSpecs from monitor
objects.

3) The is a lot of common logic for any type of datasource:
e.g.
- How to normalize DNS Names.
- When to upgrade to a Wildcard CertSpec.
- How to check PEM-cert validity
etc.
A lot of the code covering these tasks, have been moved to the module
"common".

fixes #18