This fixes issues with monitoring of both k8s- and file specs
The main issue was not having a consistent identifier for each
certificate, causing certs to be re-issued even though existing certs
were still valid.

The Monitor loop now constructs the CertSpec as early as possible,
included in the CertSpec is a "CertIdentifier". "CertIdentifier" is the key that maps a
Request for a cert (CertSpec) to an actual cert (either in the
filesystem or present in k8s storage).

Futhermore, this commit restrict certain combinations of names included
in certspecs until these (one day) is better supported. Mainly, if
faytheconfig.issuewildcardcerts = true, then it's not possible to
either:

1) have *.somedomain.com (wildcard hosts) as cn

2) have any sans at all. Sans are (so far) only supported, then issue
   wildcard auto-magic is disabled.